update D4xx to be slower
previously D4xx slide speed matched FAxy, but given that D4 effect values can be over 16x larger than FAxy values, it lacked in resolution, with most of its range consumed in not-so-useful values.

now D410 matches FA01/FA10, so it's familiar and has the same range as
FAxy, while still having 4x the resolution of 0Axy. this is accomplished
by having D4 give 4x the speed of D3.