add handy preset buttons to "scale..." menu
i typically dread using "scale..." because it's inconvenient (i'm in "mouse mode" and don't want to have to select and type, and it takes a lot of clicks to use the +/- buttons).  at the same time, i'm generally not looking for precision.  hence, adding some buttons for scaling common multipliers.