ffmpeg export: implement Subprocess::wait
Oh my fucking god this was so goddamn hard to implement. But it's here.

I don't know how I was so lucky but I came across this:

https://stackoverflow.com/questions/53734902/why-child-process-is-waiting-while-using-fork-and-pipes-for-stdin-from-parents

which made me realize stuff can go bad if I do not close the file
descriptors at every possible moment. In this case, I had to close as
many unrelated ends of stdinPipe as possible so the subprocess could
figure out the export was finished