MORE NOTES FOR NEW OSC BUFFER API
in your init or setFlags function,
instead of `oscBuf[i]->rate=rate;` do `oscBuf[i]->setRate(rate)`. this is necessary as it sets the appropriate needle multiplier for it to remain 65536Hz.

if your dispatch submits to the osc buffers at a slower rate (check out nes.cpp for an example), do not divide the rate anymore. it will work regardless.

this ends the notes (hopefully)